+++ CHANGES.md
@@ -13,7 +13,7 @@
 
 - [ADD] 音声ルート変更イベントとして `SoraHandlers.onChangeAudioRoute` を追加する
   - `RTCAudioSessionDelegate.audioSessionDidChangeRoute` を利用してコールバックする
-  - コールバック引数として `RTCAudioSession` と `previousRoute` を渡す
+  - コールバック引数として `RTCAudioSession`, `reason`, `previousRoute` を渡す
   - 音声ルートの状態などの評価はコールバック内で `RTCAudioSession.currentRoute` を参照する
   - @t-miya
 - [ADD] iOS 端末画面をキャプチャして配信する ScreenCapture を追加する